+++ RELEASES.md
@@ -40,3 +40,18 @@
 ### 3.0.11
 - Update toolkit version to 3.0.5 to include `@actions/core@^1.10.0`
 - Update `@actions/cache` to use updated `saveState` and `setOutput` functions from `@actions/core@^1.10.0`
+
+### 3.1.0-beta.1
+- Update `@actions/cache` on windows to use gnu tar and zstd by default and fallback to bsdtar and zstd if gnu tar is not available. ([issue](https://github.com/actions/cache/issues/984))
+
+### 3.1.0-beta.2
+- Added support for fallback to gzip to restore old caches on windows.
+
+### 3.1.0-beta.3
+- Bug fixes for bsdtar fallback if gnutar not available and gzip fallback if cache saved using old cache action on windows.
+
+### 3.2.0-beta.1
+- Added two new actions - [restore](restore/action.yml) and [save](save/action.yml) for granular control on cache.
+
+### 3.2.0
+- Released the two new actions - [restore](restore/action.yml) and [save](save/action.yml) for granular control on cache